Update 'swr' dep to 1.0.0
This fixes issue #2314.

Hooks of `swr`, such as `useSWR` or `useSWRInfinite`, now
require a `fetcher` function. In a previous version
of `swr`, this was not always the case.

Either way, to fix these issues, we pass a default fetcher
function to these hooks when we do not require a specific
way to handle the response.