Update 'swr' dep to 1.0.0 (#2342)
* Update 'swr' dep to 1.0.0

This fixes issue #2314.

Hooks of `swr`, such as `useSWR` or `useSWRInfinite`, now
require a `fetcher` function. In a previous version
of `swr`, this was not always the case.

Either way, to fix these issues, we pass a default fetcher
function to these hooks when we do not require a specific
way to handle the response.

* Add global fetcher for SWR

To avoid specifying the same default fetcher,
we add a global config as a SWRConfig component
in the root of the app.