Adapt syntax highlighting on dark theme
All code blocks that the Telescope back-end sends us are annotated with
highlight.js classes, so the front-end has to provide the stylesheet
that defines these.

Since the stylesheet is global, we have to link both stylesheets for
light and dark theme, and disable either depending on the current
theme.